Update genome_vis.py
When opening the txt file, first line (which normally contains genome information as seen in h_cinaedi.txt) is skipped.

dimensions(): If seq_length is a perfect square, array dimensions will be square root. If not, the closest factor pair. If seq_length is prime, final nucleotide of seq is removed and function runs again.